backend/requirements.txt
fastapi
uvicorn[standard]
motor
redis
celery
pydantic
pydantic-settings
python-dotenv
httpx
twelvelabs
google-generativeai
browser-use
langchain-google-genai
playwright
pytest
pytest-asyncio
httpx
